refactor(server): extract legacy routes into routes/legacyRoutes.ts behind the flag (DS-3)
routes.ts drops from ~2,220 lines to ~120: the CityCatalyst OAuth flow,
sample-mode init, authed city/project CRUD, on-demand geospatial fetchers,
impact-model LLM endpoints, and the 11-endpoint project state machine
(plus their bottom helpers) move verbatim to routes/legacyRoutes.ts,
registered via DYNAMIC import when ENABLE_LEGACY_ROUTES != '0' — the
testRoutes pattern. Flag off now means the routes don't exist AND none of
the legacy services load.

- osm-search (Nominatim) extracted FIRST and kept in routes.ts: it's the
  workshop map's geocoding — the one workshop-reachable endpoint in the
  monolith (audit's explicit carve-out). Verified self-contained.
- agentRoutes + conceptNoteRoutes registration moves inside
  registerLegacyRoutes (same flag as before).
- DS-2's runtime middleware is superseded by registration gating: flag-off
  requests now fall through to the SPA like any unknown route.
- openaiClient goes LAZY (construct on first use): its module-load
  'new OpenAI' crashed the workshop server at boot without a key via the
  uploadRoutes → fileExtract chain. A bare local boot with ZERO legacy env
  (no OAuth ids, no OpenAI keys) now works when the flag is off.
- Registration order preserved: legacy first, workshop route files after —
  exactly as when the blocks were inline (tile-proxy shadowing lesson).

Verified: FULL chromium e2e suite 59 passed / 3 self-skipped / 0 failed;
boot smoke both states (flag off: zero legacy env boots, workshop
endpoints normal; flag on: legacy handlers respond exactly as before —
project state, concept-note, agent conversations, sample init). tsc:
server clean, only the 2 pre-existing client errors remain.
